fix(org): always scroll back to column 0
Fix the old and annoying problem of not coming back to column 0 when the current
is wider than the window, we go to end of the line (hence scrolling horizontally
to the end of the line) and we go back to the beginning of the line.

Prior of this change the buffer scrolled left until the first visible character
of the current line... which was annoying.